+++ mex/editor/rules/state.py
@@ -1,5 +1,6 @@
 import copy
-from typing import TYPE_CHECKING, cast
+from collections.abc import AsyncGenerator, Generator
+from typing import TYPE_CHECKING, Literal, cast
 
 import reflex as rx
 from pydantic import ValidationError
@@ -44,7 +45,7 @@ from mex.editor.transform import (
 from mex.editor.utils import resolve_editor_value, resolve_identifier
 
 if TYPE_CHECKING:
-    from collections.abc import AsyncGenerator, Generator, Sequence
+    from collections.abc import AsyncGenerator, Generator
 
     from reflex.event import EventSpec
 
@@ -57,6 +58,7 @@ class RuleState(State, LocalStorageMixinState):
     _api_fields: list[EditorField] = []
 
     is_submitting: bool = False
+    delete_reset_mode: None | Literal["delete", "reset"] = None
     item_title: list[EditorValue] = []
     fields: list[EditorField] = []
     stem_type: str | None = None
@@ -102,7 +104,7 @@ class RuleState(State, LocalStorageMixinState):
             self.delete_draft(self.draft_id)  # type: ignore[operator]
 
     @rx.var(cache=True, deps=["fields", "current_locale"])
-    def translated_fields(self) -> Sequence[FieldTranslation]:
+    def translated_fields(self) -> list[FieldTranslation]:
         """Compute the translated fields based on fields and current_locale.
 
         Returns:
@@ -137,6 +139,19 @@ class RuleState(State, LocalStorageMixinState):
                     if editor_value.identifier and not editor_value.text:
                         async with self:
                             await resolve_editor_value(editor_value)
+
+    @classmethod
+    def _contains_any_rule(
+        cls, rule_set: AnyRuleSetResponse | AnyRuleSetRequest
+    ) -> bool:
+        """Check if a rule set response contains any rule."""
+        return any(
+            [
+                rule_set.additive.model_dump(exclude_defaults=True),
+                rule_set.subtractive.model_dump(exclude_defaults=True),
+                rule_set.preventive.model_dump(exclude_defaults=True),
+            ]
+        )
 
     def _get_extracted_items(self) -> list[AnyExtractedModel]:
         """Get the list of extracted items the rules should apply to."""
@@ -175,6 +190,7 @@ class RuleState(State, LocalStorageMixinState):
     @rx.event
     def refresh(self) -> Generator[EventSpec]:
         """Refresh the edit or create page."""
+        self.delete_reset_mode = None
         self.fields.clear()
         self.validation_messages.clear()
 
@@ -211,6 +227,9 @@ class RuleState(State, LocalStorageMixinState):
                 validation=Validation.LENIENT,
             )
             self.item_title = transform_models_to_title([preview])
+
+        if rule_set and self._contains_any_rule(rule_set):
+            self.delete_reset_mode = "reset" if extracted_items else "delete"
 
         loaded_fields = transform_models_to_fields(
             extracted_items,