+++ tests/edit/test_main.py
@@ -1,14 +1,20 @@
 import re
+from typing import Literal
 
 import pytest
 from playwright.sync_api import Page, expect
 
 from mex.common.backend_api.connector import BackendApiConnector
-from mex.common.fields import MERGEABLE_FIELDS_BY_CLASS_NAME
+from mex.common.fields import (
+    MERGEABLE_FIELDS_BY_CLASS_NAME,
+    REQUIRED_FIELDS_BY_CLASS_NAME,
+)
 from mex.common.models import (
     ActivityRuleSetRequest,
     AdditiveActivity,
+    AdditiveContactPoint,
     AnyExtractedModel,
+    ContactPointRuleSetRequest,
     ExtractedActivity,
     ExtractedOrganizationalUnit,
     ExtractedPrimarySource,
@@ -16,7 +22,6 @@ from mex.common.models import (
 )
 from mex.common.transform import ensure_prefix
 from mex.common.types import ActivityType, Identifier, Text, TextLanguage, Theme
-from mex.editor.fields import REQUIRED_FIELDS_BY_CLASS_NAME
 from mex.editor.rules.transform import get_required_mergeable_field_names
 
 
@@ -32,6 +37,79 @@ def edit_page(
     page_body = page.get_by_test_id("page-body")
     expect(page_body).to_be_visible()
     return page
+
+
+@pytest.fixture
+def load_delete_reset_data(
+    dummy_data: list[AnyExtractedModel],
+) -> dict[Literal["delete", "reset"] | None, str]:
+    connector = BackendApiConnector.get()
+    connector.ingest(dummy_data)
+
+    delete_resp = connector.create_rule_set(
+        ContactPointRuleSetRequest(
+            additive=AdditiveContactPoint(email=["DELETE_ME@test.de"])
+        )
+    )
+    cp1: AnyExtractedModel = next(
+        x for x in dummy_data if x.identifierInPrimarySource == "cp-1"
+    )
+    connector.update_rule_set(
+        cp1.stableTargetId,
+        ContactPointRuleSetRequest(
+            additive=AdditiveContactPoint(email=["another_email@test.com"])
+        ),
+    )
+
+    ou1: AnyExtractedModel = next(
+        x for x in dummy_data if x.identifierInPrimarySource == "ou-1"
+    )
+
+    return {
+        "delete": delete_resp.stableTargetId,
+        "reset": cp1.stableTargetId,
+        None: ou1.stableTargetId,
+    }
+
+
+@pytest.mark.integration
+def test_edit_page_delete_reset_button(
+    base_url: str,
+    writer_user_page: Page,
+    load_delete_reset_data: dict[Literal["delete", "reset"] | None, str],
+) -> None:
+    page = writer_user_page
+
+    def _navigate(stable_id: str) -> None:
+        page.goto(f"{base_url}/item/{stable_id}")
+        page_body = page.get_by_test_id("page-body")
+        expect(page_body).to_be_visible()
+
+    _navigate(load_delete_reset_data[None])
+    expect(page.get_by_test_id("delete-reset-button")).not_to_be_visible()
+
+    reset_id = load_delete_reset_data["reset"]
+    _navigate(reset_id)
+    button = page.get_by_test_id("delete-reset-button")
+    expect(button).to_be_visible()
+    button.click()
+    page.screenshot(
+        path="tests_edit_test_main-test_edit_page_delete_reset_button-reset_clicked.png"
+    )
+    page.wait_for_url(f"{base_url}/item/{reset_id}")
+    expect(page.locator(".editor-toast")).to_be_visible()
+    expect(page.get_by_test_id("delete-reset-button")).not_to_be_visible()
+
+    delete_id = load_delete_reset_data["delete"]
+    _navigate(delete_id)
+    button = page.get_by_test_id("delete-reset-button")
+    expect(button).to_be_visible()
+    button.click()
+    page.screenshot(
+        path="tests_edit_test_main-test_edit_page_delete_reset_button-delete_clicked.png"
+    )
+    page.wait_for_url(f"{base_url}/")
+    expect(page.locator(".editor-toast")).to_be_visible()
 
 
 @pytest.mark.integration